style: добавления checkstyle

+++ src/main/java/game/view/renderers/GridRenderer.java
@@ -3,7 +3,7 @@ package game.view.renderers;
 import javafx.scene.canvas.GraphicsContext;
 import javafx.scene.paint.Color;
 
-public class GridRenderer implements Renderer{
+public class GridRenderer implements Renderer {
     private final Color color;
     private final int canvasWidth;
     private final int canvasHeight;
@@ -20,11 +20,11 @@ public class GridRenderer implements Renderer{
         gc.setStroke(color);
         gc.setLineWidth(0.5);
 
-        for (int x = 0; x<= canvasWidth; x+= (int) tileSize) {
+        for (int x = 0; x <= canvasWidth; x += tileSize) {
             gc.strokeLine(x, 0, x, canvasHeight);
         }
 
-        for (int y = 0; y<= canvasHeight; y+= (int) tileSize) {
+        for (int y = 0; y <= canvasHeight; y += tileSize) {
             gc.strokeLine(0, y, canvasWidth, y);
         }
     }
